Python/requirements.txt
# xyzTableServer dependencies
# Install with: pip install -r requirements.txt

pyserial
colorama
python-dotenv

